Arrows now register with transitions (important)
Okay, listen up:

You know how we used to set inputs and outputs in the transition? We're
not doing this anymore. We have to set the arrows anyway, so they can
just tell the transitions that they belong to them. Saves some work.

But - wouldn't it save even more work to never create the arrows
manually at all, and instead opt to create arrows utomatically from the
transitions?

Listen - you're right. That _would_ be easier, ebcause initializing the
arrows is a bit of a pain. BUT - there's a feature coming.
Multiplicities of inputs/outputs. This will be very easy to implement
and use this way around. It would be a pain the other way. That's why
were doing this. Cool?

Cool.